+++ addons/suppress/RscTitles.hpp
@@ -1,0 +1,4 @@
+class RscTitles {
+    #include "\ORA\BNA_KC\addons\core\ui\BaseControls.hpp"
+    #include "ui\overlay.hpp"
+};
