+++ addons/suppress/XEH_PREP.hpp
@@ -1,4 +1,3 @@
-
 PREP(displayLoad);
 PREP(displayUnload);
 PREP(handlePlayerChanged);
