+++ addons/suppress/XEH_postInit.sqf
@@ -1,11 +1,5 @@
 #include "script_component.hpp"
 
-if !(hasInterface) exitWith {};
-
-["CBA_settingsInitialized", {
-  // Create overlay and hide it
-  0 spawn FUNC(initOverlay);
-
-  // Add unit EH and tigger it for current unit
-  ["unit", {_this call FUNC(handlePlayerChanged)}, true] call CBA_fnc_addPlayerEventHandler;
-}] call CBA_fnc_addEventHandler;
+if (hasInterface) then {
+    call COMPILE_SCRIPT(XEH_postInitClient);
+};
