+++ addons/suppress/functions/fnc_displayLoad.sqf
@@ -1,14 +1,14 @@
-#include "script_component.hpp"
+#include "..\script_component.hpp"
 /*
  * Author: SzwedzikPL
  * Handles displayLoad event for interrupting displays
  */
 
- params ["_displayClass"];
+params ["_displayClass"];
 
- GVAR(interruptingDisplays) pushBackUnique _displayClass;
- GVAR(displayInterrupt) = (GVAR(interruptingDisplays) isNotEqualTo []);
+GVAR(interruptingDisplays) pushBackUnique _displayClass;
+GVAR(displayInterrupt) = (GVAR(interruptingDisplays) isNotEqualTo []);
 
- if (GVAR(displayInterrupt)) then {
-   0 call FUNC(resetEffects);
- };
+if (GVAR(displayInterrupt)) then {
+    0 call FUNC(resetEffects);
+};
