+++ addons/suppress/functions/fnc_displayUnload.sqf
@@ -1,4 +1,4 @@
-#include "script_component.hpp"
+#include "..\script_component.hpp"
 /*
  * Author: SzwedzikPL
  * Handles displayUnload event for interrupting displays
@@ -9,7 +9,7 @@ params ["_displayClass"];
 private _index = GVAR(interruptingDisplays) find _displayClass;
 
 if (_index isNotEqualTo -1) then {
-  GVAR(interruptingDisplays) deleteAt _index;
+    GVAR(interruptingDisplays) deleteAt _index;
 };
 
 GVAR(displayInterrupt) = (GVAR(interruptingDisplays) isNotEqualTo []);
