+++ addons/suppress/functions/fnc_handleSuppressed.sqf
@@ -1,4 +1,4 @@
-#include "script_component.hpp"
+#include "..\script_component.hpp"
 /*
  * Author: SzwedzikPL
  * Handles suppressed EH
@@ -7,18 +7,18 @@
 params ["_unit", "_distance", "_shooter", "", "_ammoObject"];
 
 if (
-  // Exit if EH is triggered for remote unit - should not happen
-  !(local _unit) ||
-  // Exit if unit is dead (can be spectator)
-  {!(alive _unit)} ||
-  // Exit if some display is interrpting (arsenal, spectator, curator, etc.)
-  {GVAR(displayInterrupt)} ||
-  // Exit if unit in vehicle
-  {!(isNull (objectParent _unit))} ||
-  // Exit if unit is unconscious
-  {_unit getVariable ["ACE_isUnconscious", false]} ||
-  // Exit if shooter is closer than min distance
-  {!(isNull _shooter) && {((_unit distance _shooter) < GVAR(shooterMinDistance))}}
+    // Exit if EH is triggered for remote unit - should not happen
+    !(local _unit) ||
+    // Exit if unit is dead (can be spectator)
+    {!(alive _unit)} ||
+    // Exit if some display is interrpting (arsenal, spectator, curator, etc.)
+    {GVAR(displayInterrupt)} ||
+    // Exit if unit in vehicle
+    {!(isNull (objectParent _unit))} ||
+    // Exit if unit is unconscious
+    {_unit getVariable ["ACE_isUnconscious", false]} ||
+    // Exit if shooter is closer than min distance
+    {!(isNull _shooter) && {((_unit distance _shooter) < GVAR(shooterMinDistance))}}
 ) exitWith {};
 
 private _newSuppress = linearConversion [GVAR(projectileMaxDistance), 1, _distance, 0, 1, true];
@@ -41,7 +41,7 @@ private _suppress = (_newSuppress + _curSuppress) min 1;
 _overlayCtrl ctrlSetFade (1 - _suppress);
 _overlayCtrl ctrlCommit 0.07;
 _overlayCtrl spawn {
-  sleep 0.07;
-  _this ctrlSetFade 1;
-  _this ctrlCommit GVAR(overlayFadeoutTime);
+    sleep 0.07;
+    _this ctrlSetFade 1;
+    _this ctrlCommit GVAR(overlayFadeoutTime);
 };
