+++ addons/suppress/functions/fnc_resetEffects.sqf
@@ -1,4 +1,4 @@
-#include "script_component.hpp"
+#include "..\script_component.hpp"
 /*
  * Author: SzwedzikPL
  * Resets suppression effect
