--- addons/suppress/functions/script_component.hpp
@@ -1,1 +1,0 @@
-#include "..\script_component.hpp"
