+++ addons/suppress/initSettings.inc.sqf
